Rework some devstack things

+++ env-scripts/devstack.bash
@@ -5,24 +5,27 @@ set -e
 cd $HOME
 
 apt-get update
-apt-get install git vim-nox tmux python-setuptools zsh
+apt-get install -y git vim-nox tmux python-setuptools zsh
 
 easy_install pip
 pip install virtualenv virtualenvwrapper tox
 
 USER_HOME=/home/jim
-useradd -d $USER_HOME -m -s /bin/zsh jim
+if [[ ! -e "$USER_HOME" ]]; then
+    useradd -d $USER_HOME -m -s /bin/zsh jim
 
-mkdir $USER_HOME/.ssh
-cp $HOME/.ssh/authorized_keys $USER_HOME/.ssh/
-chown -R jim:jim $USER_HOME/.ssh
+    mkdir $USER_HOME/.ssh
+    cp $HOME/.ssh/authorized_keys $USER_HOME/.ssh/
+    chown -R jim:jim $USER_HOME/.ssh
+fi
 
-git clone git@github.com:openstack-dev/devstack
-cd devstack
-STACK_USER=jim
-DEST=$USER_HOME
-bash tools/create-stack-user.sh
-cd -
+if [[ ! -e devstack ]]; then
+    git clone git@github.com:openstack-dev/devstack
+    cd devstack
+    STACK_USER=jim
+    DEST=$USER_HOME
+    bash tools/create-stack-user.sh
+    cd -
+fi
 
-wget 'https://raw.githubusercontent.com/jimrollenhagen/dotfiles-redux/master/env-scripts/devstack-user.bash'
-su jim -c 'bash devstack-user.bash'
+wget 'https://raw.githubusercontent.com/jimrollenhagen/dotfiles-redux/master/env-scripts/devstack-user.bash' -O $USER_HOME/devstack-user.bash
